refactor: Sidebar cleanup
