+++ src/layouts/Sidebar/index.ts
@@ -1,3 +1,2 @@
-export { default as Navbar } from "@/layouts/Sidebar/components/Navbar";
-export { NavbarCategories } from "@/layouts/Sidebar/constants/NavbarCategories";
-export { default as getNavbarIcon } from "@/layouts/Sidebar/utils/getNavbarIcon";
+export { default as Navbar } from "@/layouts/Sidebar/components/Sidebar";
+export { SidebarCategories } from "@/layouts/Sidebar/constants/SidebarCategories";
